make scad2stl independent callable script
