+++ api/sp2_api.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 from typing import Union
 from pathlib import Path
-from packaging import version
+
 import copy
 from math import sin, sqrt, ceil
 from solid2 import cube, sphere, polygon, text, cylinder, import_, import_scad
@@ -16,47 +16,9 @@ sys.path.append(os.path.join(os.path.dirname(__file__), '../'))
 from api.pylele_api import ShapeAPI, Shape, Fidelity, Implementation
 from api.pylele_api_constants import DEFAULT_TEST_DIR
 from api.pylele_utils import ensureFileExtn, descreteBezierChain, superGradient, encureClosed2DPath, stl2bin
+from api.scad2stl import scad2stl
 
 FIDELITY_K = 4
-OPENSCAD='openscad'
-
-def openscad_version():
-    """ Returns openscad version """
-    tmplog='log.txt'
-    cmdstr = f'{OPENSCAD} -v 2>&1 | cat > {tmplog}'
-    os.system(cmdstr)
-
-    assert os.path.isfile(tmplog), f'ERROR: file {tmplog} does not exist!'
-
-    with open(tmplog, encoding='utf-8') as f:
-        lines = f.readlines()
-
-    # print(f'<{lines}>')
-    ans = lines[0].split()
-    assert len(ans)==3, f'Missing arguments in openscad version output <{ans}>'
-
-    ver=ans[2]
-    # print(ver)
-
-    # remove temporary file
-    os.system(f'rm {tmplog}')
-
-    return ver
-
-def openscad_manifold_ok() -> bool:
-    """ check manifold available """
-    # https://github.com/openscad/openscad/issues/391#issuecomment-1718145488
-    ver = openscad_version()
-    # print(ver)
-    if version.parse(ver) > version.parse("2023.09"):
-        return True
-    return False
-
-def openscad_manifold_opt() -> str:
-    """ generate manifold option enable, if available """
-    if openscad_manifold_ok():
-        return '--enable=manifold'
-    return ''
 
 class Sp2ShapeAPI(ShapeAPI):
     """
@@ -78,13 +40,7 @@ class Sp2ShapeAPI(ShapeAPI):
     def exportSTL(self, shape: Sp2Shape, path: str) -> None:
         basefname, _ = os.path.splitext(path)
         scad_file = self.exportBest(shape=shape, path=basefname)
-        fout = ensureFileExtn(basefname,'.stl')
-        manifold = openscad_manifold_opt()
-        cmdstr = f'{OPENSCAD} {manifold} -o {fout} {scad_file}'
-        os.system(cmdstr)
-
-        assert os.path.isfile(fout), f'ERROR: file {fout} does not exist!'
-        return fout
+        return scad2stl(scad_file)
         
     def exportBest(self, shape: Sp2Shape, path: Union[str, Path]) -> str:
         outdir, fname = os.path.split(path)
